Static gzip compression (gulp-gzip)
We can save decent junk of network traffic by sending gzipped data
This only needs one extra line in nginx conf
and each file needs to have its .gz version,
which we do with the help of gulp-gzip.
`gulp production` generates compressed files
but `gulp dev` does not (just to speed up the build process)

Besides js and css, we also compress a couple of larger svg's
from homepage. We store these in the repo as they are unlikely to change
but they can be gzipped by running `gulp compress-img`

This covers all the biggest assets, so we don't need dynamic gzip
compression -- i.e. this change should not incur CPU% penalty on the
server.

Also, we now copy all js libs to build/js/libs
and we also minify libs that were not minified.

This change cuts the homepage transfer from ~570kB to ~210kB