Remove AES for now.
mcrypt library that we used is deprecated in 7.1
and completely removed in 7.2.
For now, it's okay to simply store passwords as hashes,
further encryption is not a big security win anyway.

I kept the class and functions intact in case we wanted
to use them again, but we need to remember that they
currently do not encrypt/decrypt!

If we want to restore the functionality, see e.g.
https://stackoverflow.com/questions/41272257/mcrypt-is-deprecated-what-is-the-alternative
https://paragonie.com/blog/2015/05/if-you-re-typing-word-mcrypt-into-your-code-you-re-doing-it-wrong
https://paragonie.com/blog/2015/05/using-encryption-and-authentication-correctly